+++ projects/cocoapods-interactor/lib/cocoapods_interactor.rb
@@ -2,3 +2,5 @@
 
 module CocoaPodsInteractor
 end
+
+puts "hello world"