+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 
 - Add support for `excluding` parameter to `FileList` [#3773](https://github.com/tuist/tuist/pull/3773) by [@pavel-trafimuk](https://github.com/pavel-trafimuk)
 
+### Fixed
+
+- Fix default template to work with `tvos` platform [#3759](https://github.com/tuist/tuist/pull/3759) by [@ezraberch](https://github.com/ezraberch)
+
 ## 2.3.2 - Discoteque
 
 ### Fixed
@@ -16,6 +20,10 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 ### Changed
 - Update `swiftlint` to version `0.10.1` [#3744](https://github.com/tuist/tuist/pull/3744) by [@pepibumur](https://github.com/pepibumur)
 - Update `xcprettify` to version `0.45.0` [#3744](https://github.com/tuist/tuist/pull/3744) by [@pepibumur](https://github.com/pepibumur)
+
+### Added
+
+- Add `uiTests` target support for `tvOS`. [#3756](https://github.com/tuist/tuist/pull/3756) by [@sujata23](https://github.com/sujata23)
 
 ## 2.3.1 - Avantgarde